+++ app/src/main/java/com/example/bmi_calculator/MyAdapter.kt
@@ -1,13 +1,11 @@
 package com.example.bmi_calculator
 
 import android.content.Context
-import android.opengl.ETC1.getHeight
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ArrayAdapter
 import android.widget.TextView
-import androidx.core.content.ContextCompat
 
 class MyAdapter(context: Context, data: List<String>) :
     ArrayAdapter<String>(context, 0, data) {
@@ -27,7 +25,7 @@ class MyAdapter(context: Context, data: List<String>) :
         val item = getItem(position)
 
         bmi.text = getBmi(item)
-        val currentColor = getColor(context, bmi.text.toString().toFloat())
+        val currentColor = getColor(context, bmi.text.toString().toDouble())
         bmi.setTextColor(currentColor)
 
         dateRow.text = getDate(item)
